server/requirements.txt
# Gamepad server core
vgamepad
pyautogui
keyboard

# PC audio streaming (WASAPI loopback — Windows)
pyaudiowpatch
sounddevice
numpy

# PC screen streaming
mss
Pillow
